+++ lib/searchkick/reindex_v2_job.rb
@@ -2,7 +2,7 @@ module Searchkick
   class ReindexV2Job < ActiveJob::Base
     queue_as { Searchkick.queue_name }
 
-    def perform(class_name, id, method_name = nil, routing: nil, index_name: nil)
+    def perform(class_name, id, method_name = nil, routing: nil, index_name: nil, allow_missing: false)
       model = Searchkick.load_model(class_name, allow_child: true)
       index = model.searchkick_index(name: index_name)
       # use should_index? to decide whether to index (not default scope)
@@ -11,7 +11,7 @@ module Searchkick
       # but keep for now for backwards compatibility
       model = model.unscoped if model.respond_to?(:unscoped)
       items = [{id: id, routing: routing}]
-      RecordIndexer.new(index).reindex_items(model, items, method_name: method_name, single: true)
+      RecordIndexer.new(index).reindex_items(model, items, method_name: method_name, allow_missing: allow_missing, single: true)
     end
   end
 end